Refactor and Clean up
- Remove redshift emitter (We don’t need it for now)
- Refactor emitter package
- Remove example and unneeded code
- Fix lint and vet
