+++ emitter/emitter_test.go
@@ -1,4 +1,4 @@
-package s3
+package emitter_test
 
 import (
 	"fmt"
@@ -6,14 +6,15 @@ import (
 	"time"
 
 	"github.com/bmizerany/assert"
+	"github.com/telenor-digital-asia/kinesis-connectors/emitter"
 )
 
 func Test_Key(t *testing.T) {
 	d := time.Now().UTC().Format("2006/01/02")
 
-	k := Key("", "a", "b")
+	k := emitter.Key("", "a", "b")
 	assert.Equal(t, k, fmt.Sprintf("%v/a-b", d))
 
-	k = Key("prefix", "a", "b")
+	k = emitter.Key("prefix", "a", "b")
 	assert.Equal(t, k, fmt.Sprintf("prefix/%v/a-b", d))
 }
