--- emitter/redshift/manifest.go
@@ -1,10 +1,0 @@
-package redshift
-
-type Entry struct {
-	Url       string `json:"url"`
-	Mandatory bool   `json:"mandatory"`
-}
-
-type Manifest struct {
-	Entries []Entry `json:"entries"`
-}
